separate and export Toast item as stateless
