+++ src/lib/components/Toast/index.js
@@ -1,2 +1,3 @@
 export { withToastProvider } from './withToastProvider';
 export { useToast } from './useToast';
+export { default as Toast } from './Toast';
